useClock
