+++ react-hooks-basic/src/hooks/useClock.js
@@ -1,0 +1,33 @@
+import { useState, useEffect } from "react";
+
+function formatDate(date) {
+  if (!date) return "";
+
+  const hours = `0${date.getHours()}`.slice(-2);
+  const minutes = `0${date.getMinutes()}`.slice(-2);
+  const seconds = `0${date.getSeconds()}`.slice(-2);
+  return `${hours}:${minutes}:${seconds}`;
+}
+
+function useCLock(props) {
+  const [timeString, setTimeString] = useState("");
+
+  useEffect(() => {
+    const clockInterval = setInterval(() => {
+      const now = new Date();
+      // HH:mm:ss
+      const newTimeString = formatDate(now);
+      setTimeString(newTimeString);
+    }, 1000);
+
+    return () => {
+      // cleanup
+      console.log("Clock cleanup");
+      clearInterval(clockInterval);
+    };
+  }, []);
+
+  return { timeString };
+}
+
+export default useCLock;
